Fix: Transistor base and emitter pin assignments were flipped
In the inital version of the digikey footprint library that I
was using, the pin assignments were wrong on the transistor
schematic. This lead me to wiring up the emitter to where the
base was supposed to go and vice versa. As a result, backlighting
will not work on any revision of the board prior to this commit.
This commit updates the digikey library, which has the fixed
pin assignments, and re-routes the transistor pads correctly.